refactor: change hexmanagement to management
