+++ backend/docker-compose.yml
@@ -17,12 +17,12 @@ services:
     networks:
       - solontax-network
 
-  solontax-g1-hexmanagement:
+  solontax-g1-management:
     build:
       context: .
-      dockerfile: solontax-g1-hexmanagement/Dockerfile
-    container_name: solontax-g1-hexmanagement
-    image: solontax-g1-hexmanagement
+      dockerfile: solontax-g1-management/Dockerfile
+    container_name: solontax-g1-management
+    image: solontax-g1-management
     restart: always
     ports:
       - "1975:1975"
