--- backend/solontax-g1-hexmanagement/src/main/java/solontax/g1/hexmanagement/SolontaxG1HexmanagementApplication.java
@@ -1,13 +1,0 @@
-package solontax.g1.hexmanagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SolontaxG1HexmanagementApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SolontaxG1HexmanagementApplication.class, args);
-    }
-
-}
