--- backend/solontax-g1-hexmanagement/src/main/java/solontax/g1/hexmanagement/domain/port/PersonRepositoryPort.java
@@ -1,17 +1,0 @@
-package solontax.g1.hexmanagement.domain.port;
-
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.domain.Page;
-import solontax.g1.hexmanagement.common.dto.PersonQueryParams;
-import solontax.g1.hexmanagement.domain.model.Person;
-
-public interface PersonRepositoryPort {
-    Person save(Person person);
-
-    Optional<Person> findById(UUID id);
-
-    Optional<Person> findByTaxNumber(Long taxNumber);
-
-    Page<Person> search(PersonQueryParams personQueryParams);
-}
