--- backend/solontax-g1-hexmanagement/src/test/java/solontax/g1/hexmanagement/SolontaxG1HexmanagementApplicationTests.java
@@ -1,15 +1,0 @@
-package solontax.g1.hexmanagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class SolontaxG1HexmanagementApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
