+++ backend/solontax-g1-management/src/main/java/solontax/g1/management/application/dto/PersonDto.java
@@ -1,0 +1,27 @@
+package solontax.g1.management.application.dto;
+
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class PersonDto {
+    private UUID id;
+
+    private String firstName;
+
+    private String lastName;
+
+    private Long age;
+
+    private Long taxNumber;
+
+    private Long taxDebt;
+}
