+++ backend/solontax-g1-management/src/main/java/solontax/g1/management/common/dto/PersonQueryParams.java
@@ -1,0 +1,16 @@
+package solontax.g1.management.common.dto;
+
+import lombok.Data;
+
+@Data
+public class PersonQueryParams {
+    private String query;
+
+    private String sortBy = "id";
+
+    private String sortDirection = "asc";
+
+    private int size = 10;
+
+    private int offset = 0;
+}
