+++ backend/solontax-g1-management/src/main/java/solontax/g1/management/domain/model/Person.java
@@ -1,0 +1,28 @@
+package solontax.g1.management.domain.model;
+
+import java.time.LocalDate;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Person {
+    private UUID id;
+
+    private String firstName;
+
+    private String lastName;
+
+    private LocalDate dateOfBirth;
+
+    private Long taxNumber;
+
+    private Long taxDebt;
+}
