+++ backend/solontax-g1-management/src/test/java/solontax/g1/management/SolontaxG1ManagementApplicationTests.java
@@ -1,0 +1,15 @@
+package solontax.g1.management;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class SolontaxG1ManagementApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
